+++ src/MsCoreUtils.h
@@ -12,8 +12,12 @@ extern SEXP C_closest_dup_keep(SEXP, SEXP, SEXP, SEXP);
 extern SEXP C_closest_dup_closest(SEXP, SEXP, SEXP, SEXP);
 extern SEXP C_closest_dup_remove(SEXP, SEXP, SEXP, SEXP);
 
+extern SEXP C_gnps(SEXP, SEXP);
+extern SEXP C_gnps_chain_dp(SEXP, SEXP, SEXP, SEXP, SEXP, SEXP);
+
 extern SEXP C_impNeighbourAvg(SEXP, SEXP);
 
+extern SEXP C_join_gnps(SEXP, SEXP, SEXP, SEXP, SEXP, SEXP);
 extern SEXP C_join_left(SEXP, SEXP, SEXP, SEXP);
 extern SEXP C_join_right(SEXP, SEXP, SEXP, SEXP);
 extern SEXP C_join_inner(SEXP, SEXP, SEXP, SEXP);